requirements.txt
pillow==10.4.0
qrcode==7.4.2
typer==0.12.3
